Added new example and corrected bug.
A new example notebook that solves an actual
differential equation was implemented - solver2.ipynb.

While testing the example, I found a bug when there is
only one element and both ends of the domain consist
of Dirichilet boundary conditions. In this case,
nbslv = 0 and the there are problems when allocating/
solving tridiagonal systems. Now the code handles this
case well.

There is still a situation that is not handled that I will
correct next: When there are only boundary modes.