BBTri uses Tridiagonal{T}
Now BBTri uses Tridiagonal{T} for storing boundary-boundary matrices.
With this, code will be able to use BigFloats to solve linear systems.

For now this doesn't work because `add_local_matrix`, a method used
in static condensation, still uses Lapack. I will try to change this later.

The old code for BBTri is still there and it is called BBTri2.