Periodic boundary conditions working
It does hava a couple of caveats: it doesn't work for 1 or 2 elements domain.
The problem is that in those cases, the base library might try to allocate
e vectors with negative number of elements and this throws an error.

The code for BBTriP and BBSymTriP use lapack functions. I should probably
change that so that the Julia routines are used.